pyproject.toml
[project]
name = "roedermark-haushalt"
version = "0.1.0"
description = "Pipeline + UI für den Haushalt der Stadt Rödermark – von PDF zu offenen Daten"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "pdfplumber>=0.11",
    "pandas>=2.1",
    "duckdb>=1.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "pyarrow>=15.0",          # Parquet-Support
    "tqdm>=4.66",              # Fortschrittsanzeige
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3",
    "Pillow>=10.0",
]
dev = [
    "ruff>=0.5",
    "pytest>=8.0",
    "ipython>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
